src/third_party/WebKit/LayoutTests/fast/css/invalidation/toggle-style-inside-shadow-root-expected.txt

CONSOLE WARNING: line 5: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(rootDiv).backgroundColor is "rgb(255, 0, 0)"
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(rootDiv).backgroundColor is "rgba(0, 0, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
